Finished Product (#10)
* Necessary Files

Package.json, babel, gitignore, webpack, and images that are necessary to run a react web app.

* Feature/add page (#5)

* Primary Components

These two files are what make up the home page of the app.

* Menu

The setup for our menu with some test data inside of it

* Bar Page

This is our bar page with test data

* Bar Page and Background

Updated barpage with actual location and bar locations and made the background a gradient instead of a picture.

* Finished Bar Page

Finished up plugging in all of the actual data from foursquare and google.

* Final Requirements

Created our own cocktail database and put that into a list on the menu page. Also added two options to sort the list alphabetically and by popularity. Fixed the background issue. Added "home" buttons to the menu page and the bar page for convenience.

* Fixed Buttons

Added a disable feature for the buttons once the list is sorted

* Homebrew

Allows the user to select on-hand ingredients and returns the list of possible drinks to be made

* Homebrew

Allows the user to select on-hand ingredients and returns the list of possible drinks to be made

* Homebrew

Working buttons, broken logic.

* Add Feature

Almost done implementing the add feture, but just wanted to commit what I have

* Add Feature

Finished the add feature, and cleaned up the code and restructured it the right way

* Add/Delete Functionality

Add and Delete drinks from the Menu page and the drink list is stored on local storage so your changes will be saved.

* Homebrew up and running

Just make sure you click all the ingredients. Still a bit buggy.

* Add Feature

Almost done implementing the add feture, but just wanted to commit what I have

* Add Feature

Finished the add feature, and cleaned up the code and restructured it the right way

* Add/Delete Functionality

Add and Delete drinks from the Menu page and the drink list is stored on local storage so your changes will be saved.

* Modified List

Made a few modifications to the drink list and ingredient list

* Finished Edit Feature

Edit/Delete/Add are all working smoothly and the list is saved to local storage so you will always have your edits/adds and will not have the ones you deleted.

* Finishing Edits

Cleaned up code based on Jacks comments

* Finish Editing HomeBrew

Forgot to include the new Homebrew

* Fix/babel (#7)

* Babel File

Babel File

* Babel pull

babel

* Material UI Import (#8)

Added Material UI framework which required modifiying .babelrc and wepback file. used tutorials from https://github.com/callemall/material-ui/tree/master/examples/webpack-example and https://github.com/petehunt/webpack-howto

* Revert "Material UI Import" (#9)

* Classy Mixology University

Finished UI